Revert "unit test all pass"
This reverts commit 602c457e637425b425bbbe20975b7d3bf9927d46.

+++ app/checkout/biz/service/checkout_test.go
@@ -37,7 +37,7 @@ func TestCheckout_Run(t *testing.T) {
 	}
 
 	req := &checkout.CheckoutReq{
-		UserId:     uint32(3),
+		UserId:     3,
 		Email:      "hourhl@hourl.com",
 		Address:    address,
 		CreditCard: creditcard,
